+++ src/Callbacks.jl
@@ -83,7 +83,8 @@ export observe_gale_start!,
 export observe_gare_start!,
        observe_gare_step!,
        observe_gare_done!,
-       observe_gare_failed!
+       observe_gare_failed!,
+       observe_gare_metadata!
 export observe_gdre_start!,
        observe_gdre_step!,
        observe_gdre_done!
@@ -183,5 +184,6 @@ observe_gare_start!(::Any, args...) = nothing
 observe_gare_step!(::Any, args...) = nothing
 observe_gare_done!(::Any, args...) = nothing
 observe_gare_failed!(::Any) = nothing
+observe_gare_metadata!(::Any, args...) = nothing
 
 end